requirements.txt
fastapi
fastapi-sqlalchemy
pydantic
alembic
psycopg2-binary
uvicorn